[UPDATED] comments, removed unused imports,
cleaned project
