+++ sprit/model/station_data_analytics_model.py
@@ -1,20 +1,12 @@
 import sqlite3
-import sys
-
 import pandas as pd
 import numpy as np
-import os
-
 from PIL import Image
-
 from sprit.resources import helper
 
 
 class StationDataAnalyticsModel:
     def __init__(self):
-        # Determine the directory of the currently running script
-        base_dir = os.path.dirname(os.path.abspath(__file__))
-
         # Combine the base directory with the relative path to the database file
         db_path = helper.find_data_file("tk_hist.db")
 
@@ -33,17 +25,9 @@ class StationDataAnalyticsModel:
         self.average_price = "1.00"
         self.is_recommended = True
 
-        # Get the directory of the currently running script
-        base_dir = os.path.dirname(os.path.abspath(__file__))
-
         # Join the base directory with the relative path to the icons
         self.thumb_up_path = helper.find_data_file("thumb_up_green.png")
         self.thumb_down_path = helper.find_data_file("thumb_down_red.png")
-        #self.thumb_up_path = os.path.dirname(sys.executable)+"/thumb_up_green.png"
-        #self.thumb_down_path = os.path.dirname(sys.executable)+"/thumb_down_red.png"
-        print(self.thumb_up_path)
-        print(self.thumb_down_path)
-
         self.recommendation_icon = Image.open(self.thumb_up_path)
 
     def connect_to_hist_database(self):
