+++ sprit/model/stations_search_list_map_model.py
@@ -1,6 +1,5 @@
 from sprit.model.station_model import Station
 from PIL import Image, ImageTk
-import os
 from sprit.resources import helper
 
 
@@ -74,14 +73,10 @@ class StationsSearchListMapModel:
         Returns:
             An ImageTk.PhotoImage object representing the station's icon.
         """
-
-        # Get the directory of the currently running script
-        base_dir = os.path.dirname(os.path.abspath(__file__))
-
         try:
 
-            img = Image.open( helper.find_data_file(f"{station_name}.png"))
-        except FileNotFoundError as e:
+            img = Image.open(helper.find_data_file(f"{station_name}.png"))
+        except FileNotFoundError:
             # If specific station icon is not found, use a default icon
             helper.find_data_file(f"gas_station_icon.png")
             img = Image.open(helper.find_data_file(f"gas_station_icon.png"))
