+++ sprit/resources/helper.py
@@ -1,14 +1,14 @@
 import sys
 import os
 
+#Source: https://cx-freeze.readthedocs.io/en/latest/faq.html#using-data-files
 
 def find_data_file(filename):
     if getattr(sys, "frozen", False):
         # The application is frozen
-        datadir = os.path.dirname(sys.executable)
+        datadir = os.path.dirname(sys.executable)+"/resources/"
     else:
         # The application is not frozen
         # Change this bit to match where you store your data files:
         datadir = os.path.dirname(__file__)
-        print(os.path.join(datadir, filename))
     return os.path.join(datadir, filename)
